Change damage values to doubles

+++ text-game/src/me/cs158/tag/item/BroadSword.java
@@ -4,7 +4,7 @@ public class BroadSword extends MeleeWeapon {
 	
 	private static final Items ITEM = Items.BROADSWORD;
 
-	public BroadSword(float minDamage, float maxDamage) {
+	public BroadSword(double minDamage, double maxDamage) {
 		super(ITEM, minDamage, maxDamage);
 	}
