+++ text-game/src/me/cs158/tag/item/MeleeWeapon.java
@@ -2,7 +2,7 @@ package me.cs158.tag.item;
 
 public abstract class MeleeWeapon extends Weapon {
 	
-	public MeleeWeapon(Items item, float minDamage, float maxDamage) {
+	public MeleeWeapon(Items item, double minDamage, double maxDamage) {
 		super(item, minDamage, maxDamage);
 	}
